api/src/main/java/com/github/gustavoflor/ordo/api/security/service/UserDetailsServiceImpl.java
package com.github.gustavoflor.ordo.api.security.service;

import com.github.gustavoflor.ordo.core.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.ArrayList;

@Service
@RequiredArgsConstructor
public class UserDetailsServiceImpl implements UserDetailsService {
    private final UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(final String username) throws UsernameNotFoundException {
        return userRepository.findByEmail(username)
            .map(user -> new User(username, user.getPassword(), new ArrayList<>()))
            .orElseThrow(() -> new UsernameNotFoundException("Ops... user not found!"));
    }
}
